+++ core/animation/css_keyframe_manager.h
@@ -57,6 +57,9 @@ class CSSKeyframeManager : public AnimationDelegate {
   CSSKeyframeManager(tasm::Element* element);
   ~CSSKeyframeManager() = default;
 
+  void AddAnimationDataAndPlay(
+      base::Vector<starlight::AnimationData>& anim_data);
+
   void SetAnimationDataAndPlay(
       base::Vector<starlight::AnimationData>& anim_data);
 
@@ -100,9 +103,12 @@ class CSSKeyframeManager : public AnimationDelegate {
   std::shared_ptr<Animation> CreateAnimation(starlight::AnimationData& data);
 
   base::InlineVector<starlight::AnimationData, 1> animation_data_;
-  // animations of css style
+  // The collection of animations running on the current element.
   std::unordered_map<base::String, std::shared_ptr<Animation>> animations_map_;
-  // save active animations for play and clear
+  // The collection of animations that no need to update states during the diff.
+  std::unordered_map<base::String, std::shared_ptr<Animation>>
+      temp_keep_animations_map_;
+  // The collection of animations that need to update states during the diff.
   std::unordered_map<base::String, std::shared_ptr<Animation>>
       temp_active_animations_map_;